Fix deploy doc by removing dependency on quinoa repo
TC v2017+ reverts checkouts to server instead of agent when both quinoa
and doc repos were supposed to be checked out into the same dir, which
broke the doc deploy stage, and thus no longer updated the online
doxygen pages for about a couple of weeks.

This commit fixes this problem, by removing the dependency of teh deploy
doc step on the quinoa repo , which was only needed for reading out the
sha1 for the commit message. We also switch to transferring the TC
artifacts between doc gen and doc deploy steps in tgz, which is faster
(besides not having checkout the quinoa repo at all).